# public/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TaskHub API",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "CreateTask": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "example": "タスクタイトル"
          },
          "content": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "TODO",
              "DOING",
              "DONE"
            ],
            "description": "タスクのステータス"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateTask": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "TODO",
              "DOING",
              "DONE"
            ],
            "description": "タスクのステータス"
          }
        }
      }
    },
    "parameters": {}
  },
  "paths": {
    "/api/tasks": {
      "get": {
        "responses": {
          "200": {
            "description": "タスク一覧取得"
          }
        }
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "作成成功"
          }
        }
      }
    },
    "/api/tasks/{id}": {
      "put": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "削除成功"
          }
        }
      }
    }
  }
}